+++ src/components/button/button.tsx
@@ -6,12 +6,15 @@ import { buttonStyled, BtnCssVars } from './button.styled'
 import { Progress } from '../progress'
 import { Transition } from '../transition'
 
-export type ButtonType = 'primary' | 'default' | 'warning' | 'danger' | 'link'
+export type ButtonType = 'primary' | 'default' | 'warning' | 'danger' | 'text'
 export type ButtonHtmlType = 'button' | 'submit' | 'reset'
 
-export interface ButtonProps extends BaseComponentType<BtnCssVars> {
+export interface ButtonProps
+  extends BaseComponentType<BtnCssVars>,
+    Omit<React.HTMLAttributes<HTMLButtonElement>, 'className' | 'onClick' | 'type'> {
   /** 按钮类型 */
   type?: ButtonType
+  hoverType?: ButtonType
   /** 按钮宽度 */
   width?: number | string
   /** 按钮高度 */
@@ -22,6 +25,10 @@ export interface ButtonProps extends BaseComponentType<BtnCssVars> {
   block?: boolean
   /** 虚线边框 */
   dashed?: boolean
+  /** 圆角 */
+  radius?: number | string
+  /** 圆形按钮 */
+  round?: boolean
   /** 字体大小 */
   fontSize?: number | string
   /** 幽灵按钮 */
@@ -48,6 +55,7 @@ export const Button: React.FC<ButtonProps> = ({
   cssVars,
   style,
   type = 'default',
+  hoverType,
   width,
   height,
   paddingX,
@@ -61,20 +69,30 @@ export const Button: React.FC<ButtonProps> = ({
   disabled,
   autoLoading = true,
   autoFocus = true,
+  radius,
+  round,
   onClick,
+  onMouseEnter,
+  onMouseLeave,
   htmlType = 'button', // 默认值设为 'button'，如果不指定则在 form 中的所有 button 元素默认为 'submit'
   ...wrapProps
 }) => {
   const btnRef = React.useRef<HTMLButtonElement>(null)
   const [insideLoading, setInsideLoading] = React.useState(false)
+  const [isHover, setIsHover] = React.useState(false)
   const isLoading = loading || insideLoading
   let isDisabled = disabled || isLoading
+
+  const btnTypeClassName = React.useMemo(() => {
+    if (hoverType && isHover) return `btn-${hoverType}`
+    return `btn-${type}`
+  }, [isHover, type, hoverType])
 
   const btnClassName = useTaomuClassName(
     'btn',
     'flex center flex-none',
     block ? 'flex' : 'flex-inline',
-    `btn-${type}`,
+    btnTypeClassName,
     {
       'show-outline': showOutline,
       'btn-ghost': ghost,
@@ -85,7 +103,14 @@ export const Button: React.FC<ButtonProps> = ({
     className
   )
   const btnStyle = useInlineStyle(
-    { btnWidth: block ? '100%' : width, btnHeight: height, btnPaddingX: paddingX, btnFontSize: fontSize, ...cssVars },
+    {
+      btnWidth: block ? '100%' : width,
+      btnHeight: height,
+      btnPaddingX: paddingX,
+      btnFontSize: fontSize,
+      btnRadius: round ? 45 : radius,
+      ...cssVars,
+    },
     style
   )
 
@@ -110,6 +135,28 @@ export const Button: React.FC<ButtonProps> = ({
     }
   }
 
+  function handleMouseEnter(e: React.MouseEvent<HTMLButtonElement>) {
+    onMouseEnter?.(e)
+    if (isDisabled) {
+      e.preventDefault()
+      return
+    }
+    if (hoverType) {
+      setIsHover(true)
+    }
+  }
+
+  function handleMouseLeave(e: React.MouseEvent<HTMLButtonElement>) {
+    onMouseLeave?.(e)
+    if (isDisabled) {
+      e.preventDefault()
+      return
+    }
+    if (hoverType) {
+      setIsHover(false)
+    }
+  }
+
   return (
     <button
       ref={btnRef}
@@ -118,6 +165,8 @@ export const Button: React.FC<ButtonProps> = ({
       style={btnStyle}
       disabled={isDisabled}
       onClick={handleClick}
+      onMouseEnter={handleMouseEnter}
+      onMouseLeave={handleMouseLeave}
       type={htmlType}
       {...wrapProps}
     >